requirements.txt
adal==1.2.6
azure-common==1.1.26
azure-core==1.10.0
azure-identity==1.5.0
azure-mgmt-core==1.2.2
azure-mgmt-monitor==2.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
idna==2.10
isodate==0.6.0
msal==1.8.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.1.0
portalocker==1.7.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.3